[!!!][BUGFIX] Define correct boolean value in extension configuration
TypoScript evaluates only 0 as "false" for boolean values and interprets
each other not-empty value as truthy, thus the current extension
configuration was interpreted to be truthy. As this definitely should
not be the default value for this configuration, the default property
value "extensionServiceEnabled" has been adapted accordingly.

This is a breaking change as it switches the currently
by accident defined default value from "true" to "false" (which it
already should have been). When updating to this commit or each release
including it, make sure to adapt your extension configuration.